Allow chrony to create its own key in chrony.keys
A new parameter

* config_keys_manage

defaults to true.

If set to false then puppet will touch an already
existing chrony.keys file.

In addition chrony_password can now be specified
as the string 'unset' to signify that no password
should be added to chrony.keys file.

The aim is to allow the following configuration to
allow chrony to generate it's own key in chrony.keys
with the defaults *generatecommandkey* command
in chrony.conf and then leave that file alone.

```puppet
class{'::chrony':
  config_keys_manage => false,
  chrony_password    => 'unset',
}

Some background in
*
* https://access.redhat.com/documentation/en-US/Red_Hat_Enterprise_Linux/7/html/System_Administrators_Guide/sect-Understanding_chrony_and-its_configuration.html#sect-Security_with_chronyc